requirements.txt
python-dotenv
openai
gspread
oauth2client
nest-asyncio
matplotlib
pandas
python-telegram-bot
openpyxl
openai==1.12.0
